requirements.txt
psycopg2-binary==2.9.6
python-dotenv==1.0.0
sqlparse==0.4.4